Simplify script that generates Tokens. (#195)
Doing a little cleaning before adding support for iOS.

This removes a lot of code from `packages/thumbprint-tokens/src/index.js` by switching to Handlebars, minimizing duplicated bolierplate, and simplifying how we set up the snapshot testing.

The only change to "dist" files is that there are now newlines at the end of files. Not sure why, but it's probably a good change anyways.