Add supported platforms to each token (#198)
This makes Token `value`s an object instead of a string. The object includes the token's supported platforms and their values.

For example:

```json
{
    "name": "Blue",
    "id": "color__blue",
    "value": {
        "web": "#009fd9"
    },
    "type": "color"
}
```

We currently only support web, but a follow-up PR will add iOS support. Once that is ready, you could imagine:

```json
{
    "name": "Blue",
    "id": "color__blue",
    "value": {
        "web": "#009fd9",
        "ios": "UIColor(red:0.00, green:0.62, blue:0.85, alpha:1.0)"
    },
    "type": "color"
}
```


This change allows us to support:

* Tokens that are only be available on certain platforms. (`$wrap__max-width`, for example, would be web only.)
* Tokens that are available on two platforms but have completely different values. (`$tp-space__6` is `64px` on web but `48pt` on iOS, for example.)

This PR also makes a small change to the documentation infra so that the custom `gatsby-transformer-thumbprint-tokens` plugin is no longer needed. This is possible thanks to a [new "Schema Customization" feature](https://www.gatsbyjs.org/blog/2019-03-18-releasing-new-schema-customization/) in Gatsby.