Move prettier format check to bash script
This fixes a common pain point with developing in Thumbprint. In the current setup, the error message would appear grayed out because it is the source of the NPM script. This confused developers because it was not obvious. Moving this to a bash script allows us to hide the message unless there is an error.