Remove all deprecated tokens from Thumbprint.
There are very few in use in our codebases and they've been deprecated for a long time. Removing them will make it easier to support iOS and change some naming conventions.

Once this merges, I'll start working in `website` to remove all references to them. Tokens almost never changes, so I'm not concerned if the website migration takes a week.