Upgrade documentation packages
This updates all of the packages in the `www` folder.

While the Gatsby changes were supposed to be backwards-compatible, something related to the `gatsby-transformer-json` was causing build issues. I found a workaround which is why so many MDX files changed.

I also removed the `focus-visible` polyfill since that was causing errors. I could've fixed it instead, but it's probaly best to just remove to more closely match how `website` behaves.