Avatar documentation clean-up
This started off as a small tweak but I ended up re-writing most of the documentation.

The main change is that the page is no longer `UserAvatar` first followed by a section on `EntityAvatar`. Both are now documented together. I've also consolidated the size examples into a single example, making it easier to scan the page.