Use consistent format for package.json author data
Gatsby pulls in the `package.json` so that it can be queried for use within the documentation. It was throwing a warning about the `author` data because it was an object in some places and a string in others. This doesn't work because values within GraphQL must share the same type. Changing all of them to a string is an easy fix to remove some of the warnings.

Relates to #336.