Fix image in Chip Android documentation (#639)
Gatsby MDX and VSCode's syntax highlighter don't like spaces in image URLs. This renames the file to `chip-states.png` to remove spaces and fix the broken image.

Another option would have been to keep the file name as-is and replace the spaces with `%20` in the image path within the markdown file. I decided to rename the file since the new name is more clear.