Make it possible to `objectFit` without Image `height` prop
Users can also change the Image component height using the `height`, `className`, or `style` props.

Previously, it was only possible to use `objectFit` if `height` was also provided. This meant that, to set height via CSS _and_ use `objectFit`, users had to set the height with CSS and add a `height="100%"` prop to enable the object fit functionality.