Remove unnecessary canUseDOM check from Popover.
This was causing issues in on RR pages described in:

https://thumbtack.slack.com/archives/C0GKYQVTR/p1616022352008000

(Link is internal only.)

After creating a small reproduction, I realized that removing `canUseDOM` fixes the problem. I'm not 100% sure why, but I imagine it is related to React's client-side hydration. `canUseDOM` is always false on the server but defaults to `true` on the client. I imagine this isn't an issue with components like Tooltip because they default to being closed and require user interaction to appear.