Create a CodeExperimental component for Next.js docs
I created this component since we're moving away from MDX. It allows developers to create live React/HTML examples without needing to write it all as a string.

I decided to call it CodeExperimental because I'm gradually adopting it and it's possible that we'll run into a issue with it that becomes a blocker. So far so good though. I'll rename this
to `<Code>` and get rid of the existing `<Code>` component if this works well.