Covert the Avatar React page from MDX to TSX
This also deletes the Avatar usage page. I'll convert that to TSX in a separate PR.

Some of the code is in a messy state, but I'll clean that up once I've found the abstractions that I like. (For example, there's still an `components/mdx/` folder even though we're leaving
MDX.)