Prepare `/overview/` pages for Next.js launch
This fixes small bugs I found in the Next.js docs so that the Next.js version of these pages looks and behaves the same as the Gatsby version.

It also removes `/overview/` from the Next.js config file, telling Next.js not to proxy that URL to the Gatsby version of the site.