Migrate "Guidelines" pages to Next.js, rename URL (#910)
This moves the Guidelines pages to Next.js and also changes the URLs from `/guide/product/foo` to `/guidelines/foo`. This makes the routes consistent with the rest of the docs and it also allowed me to avoid tweaking the side-nav `isActive` logic to account for this inconsistency. I've added redirects to ensure that the old URLs continue to work. A follow-up commit will update references to the old URLs. I decided not to do it in this
commit to avoid bloating the PR.