Fix the `/components/overview/` page
We proxy some routes from Next.js to Gatsby, but we had it set as `beforeFiles`. This caused `/components/overview` to fail because the glob in one of the rewrites was catching it. Switching to `afterFiles` tells Next.js to only proxy if the page doesn't exist in Next.js.